View Models Updated

+++ SolarApp/SolarApp/SolarApp/Models/Sensor.cs
@@ -12,19 +12,19 @@ namespace SolarApp.Models
         /// <summary>
         /// Voltage extracted from sensor
         /// </summary>
-        public float ValueV { get; set; }
+        public double ValueV { get; set; }
         /// <summary>
         /// Current sensed with a connected panel
         /// </summary>
-        public float ValueI { get; set; }
+        public double ValueI { get; set; }
         /// <summary>
         /// Temp obtained by a thermistor
         /// </summary>
-        public float ValueTemp { get; set; }
+        public double ValueTemp { get; set; }
         /// <summary>
         /// Irradiance measured by piranometer.
         /// </summary>
-        public float ValueIR { get; set; }
+        public double ValueIR { get; set; }
         /// <summary>
         /// Date register by ds1307
         /// </summary>
